fix(simulation): load stateless contracts before balancer_v3 getters
A balancer_v3 pool's data getters read through the Vault, which
delegatecalls into VaultExtension. Without that implementation in the
engine every getter failed with StorageError, so the native decoder
rejected every pool. Verified against tycho-dev: pools now decode and 48
of 48 simulations reproduced their executions with zero slippage.

The loader itself moves from curve's private module to the shared VM
utils, since the stateless_contract_addr_{i} attributes it reads are a
protocol-wide reserved-attribute convention rather than anything
Curve-specific, and this is its third consumer after curve and the
generic VM adapter. Error messages lose the "curve" prefix accordingly;
behaviour is unchanged and curve's tests still pass.
